refacto: create header component
