+++ src/screens/settings/picture-delete.screen.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {
-  View, Text, Pressable, StyleSheet, ToastAndroid,
+  View, Text, Pressable, ToastAndroid,
 } from 'react-native';
 import { connect } from 'react-redux';
 import axios from 'axios';
@@ -13,6 +13,7 @@ import BxActionComponent from '../../components/bx-action.component';
 import { getUser } from '../../redux/selectors';
 import { updateUser } from '../../redux/actions';
 import ProfilePicture from '../../components/profile-picture.component';
+import BxHeader from '../../components/bx-header.component';
 
 const PictureDeleteScreen = (props: { navigation, user: AuthSubject, updateUser }) => {
   const { navigation, user, updateUser } = props;
@@ -32,57 +33,9 @@ const PictureDeleteScreen = (props: { navigation, user: AuthSubject, updateUser 
     navigation.goBack();
   };
 
-  const styles = StyleSheet.create({
-    headerContainer: {
-      paddingVertical: 20,
-      paddingHorizontal: 10,
-      borderColor: '#191919',
-      borderStyle: 'solid',
-      borderBottomWidth: 1,
-      backgroundColor: colors.background,
-    },
-    headerStyle: {
-      height: 20,
-      elevation: 0,
-      shadowOpacity: 0,
-      flexDirection: 'row',
-      justifyContent: 'flex-start',
-    },
-    previewContainer: {
-      marginTop: 20,
-    },
-    preview: {
-      flexDirection: 'row',
-      justifyContent: 'space-around',
-      marginVertical: 20,
-    },
-    colorPreview: {
-      width: 150,
-      padding: 10,
-      borderRadius: 5,
-      textAlign: 'center',
-      borderWidth: 1,
-      borderColor: '#B3B3B3',
-    },
-    settingTitle: {
-      color: colors.textColor,
-      marginLeft: 30,
-      fontWeight: '700',
-    },
-  });
-
   return (
     <>
-      <View style={styles.headerContainer}>
-        <View style={styles.headerStyle}>
-          <Pressable
-            onPress={() => navigation.pop()}
-          >
-            <BackIcon width={20} height={20} fill={colors.textColor} />
-          </Pressable>
-          <Text style={styles.settingTitle}>Remove your profile picture</Text>
-        </View>
-      </View>
+      <BxHeader text="Remove your profile picture" onPress={() => navigation.pop()} />
       <View style={{ marginHorizontal: 20, marginTop: 20 }}>
         <Text style={{ color: colors.textSystemColor, fontSize: 11, textAlign: 'center' }}>
           Your profile picture will be reset to the default picture. This operation cannot be undone, although you may change
